+++ packages/motor/test/src/controllers/track_animation_test.dart
@@ -118,7 +118,7 @@ void main() {
         AnimationStatus.forward,
         AnimationStatus.completed,
         AnimationStatus.reverse,
-        AnimationStatus.completed,
+        AnimationStatus.dismissed,
       ]);
     });